Working Funtional

+++ BankWEB/Models/BankModel.cs
@@ -10,9 +10,9 @@ namespace BankWEB.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public double InterestRate {get;set;}
-        public decimal MaximumLoan { get; set; }
-        public decimal MinimumDownPayment { get; set; }
-        public int LoanTerm { get; set; }
+        public double MaximumLoan { get; set; }
+        public double MinimumDownPayment { get; set; }
+        public uint LoanTerm { get; set; }
         public string UserId { get; set; }
     }
 }
